kafka/tx: fix aborted txns related resource leak
Redpanda needs to store a list of all ever aborted transactions.

We used to do by reading the log and then keeping the list in a snapshot.
But over time this leads to a resource leak since the list of aborted
transactions just keeps growing.

Now we break the list of the aborted transactions into segments, store
each segment in a separate file and keep the index of the segments in
memory. When Redpanda needs to get a list of aborted transactions for
a given offset range it looks up the index and then load data from disk
if data are not in the cache (we keep the newest segment in memory as
a cache).